+++ tomcat/src/main/java/jakarta/controller/ResourceFinder.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package jakarta.controller;
 
 import java.io.BufferedInputStream;
 import java.io.FileInputStream;
@@ -7,18 +7,20 @@ import java.io.InputStream;
 import java.net.URL;
 import java.util.NoSuchElementException;
 
-class ResourceFinder {
+public class ResourceFinder {
 
     private static final String RESOURCE_PATH_FORMAT = "static/%s";
     private static final String INDEX_PATH = "/";
 
-    public static boolean hasResource(String resourceName, ClassLoader classLoader) {
+    private final ClassLoader classLoader = getClass().getClassLoader();
+
+    public boolean hasResource(String resourceName) {
         URL resourcePath = classLoader.getResource(String.format(RESOURCE_PATH_FORMAT, resourceName));
 
         return resourcePath != null && !INDEX_PATH.equals(resourceName);
     }
 
-    public static byte[] readResource(String resourceName, ClassLoader classLoader) {
+    public byte[] readResource(String resourceName) {
         URL resourcePath = classLoader.getResource(String.format(RESOURCE_PATH_FORMAT, resourceName));
         if (resourcePath == null) {
             throw new NoSuchElementException("존재하지 않는 리소스입니다.");
