+++ tomcat/src/main/java/jakarta/controller/StaticResourceController.java
@@ -1,6 +1,11 @@
-package org.apache.coyote.http11;
+package jakarta.controller;
 
-class StaticResourceController extends AbstractController {
+import jakarta.http.ContentType;
+import jakarta.http.HttpRequest;
+import jakarta.http.HttpResponse;
+import jakarta.http.HttpStatus;
+
+public class StaticResourceController extends AbstractController {
 
     @Override
     public void service(HttpRequest request, HttpResponse response) throws Exception {
